backend/requirements.txt
backcall==0.1.0
decorator==4.4.1
Django==2.2.7
django-auth-ldap==2.0.0
django-cors-headers==3.2.0
djangorestframework==3.10.3
ipython==7.10.0
ipython-genutils==0.2.0
jedi==0.15.1
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
pkg-resources==0.0.0
prompt-toolkit==3.0.1
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
Pygments==2.5.2
python-ldap==3.2.0
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
traitlets==4.3.3
wcwidth==0.1.7
